Moved server.ts
