+++ src/server/routes/route.ts
@@ -1,5 +1,5 @@
 import { Response, Router } from 'express';
-import ModmailServer from '../server';
+import ModmailServer from '../../server';
 
 export default class Route {
   protected readonly name: string;
